+++ CHANGELOG.md
@@ -4,11 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/GemeenteUtrecht/bereikbaarheid-frontend/compare/v0.7.7...HEAD)
+## [Unreleased](https://github.com/GemeenteUtrecht/bereikbaarheid-frontend/compare/v0.7.8...HEAD)
+
+
+## [v0.7.8 - 2024-12-08](https://github.com/GemeenteUtrecht/bereikbaarheid-frontend/compare/v0.7.7...v0.7.8)
 
 ### Fixed
 - show user-friendly error message when API is unreachable
 - Restrictions: add tests
+- upgrade to Vite 6.x
+- updated dependencies
 
 
 ## [v0.7.7 - 2024-11-08](https://github.com/GemeenteUtrecht/bereikbaarheid-frontend/compare/v0.7.6...v0.7.7)